chore: improve bundling and shipping types of wizard-app
