+++ packages/theme-wizard-app/package.json
@@ -11,6 +11,10 @@
     "url": "https://github.com/nl-design-system/theme-wizard.git",
     "directory": "packages/theme-wizard-app"
   },
+  "sideEffects": [
+    "*.css",
+    "dist/**/*.js"
+  ],
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
@@ -18,7 +22,7 @@
     },
     "./theme.css": "./theme.css",
     "./clippy-tokens.css": "./clippy-tokens.css",
-    "./dist/components/*.js": {
+    "./components/*": {
       "import": "./dist/components/*/index.js",
       "types": "./dist/components/*/index.d.ts"
     }
@@ -32,9 +36,7 @@
     "dev": "vite build --watch --mode development",
     "types": "tsc -p tsconfig.json --emitDeclarationOnly",
     "type-check": "tsc --noEmit --project tsconfig.json",
-    "build": "pnpm run type-check && pnpm run types && pnpm run build:app && pnpm run build:components",
-    "build:app": "vite build",
-    "build:components": "vite build --config vite.components.config.ts",
+    "build": "pnpm run type-check && vite build",
     "clean": "rimraf dist/",
     "lint": "pnpm run --sequential '/^lint:.+$/'",
     "lint:css": "stylelint .",
@@ -77,7 +79,6 @@
     "@nl-design-system-community/ma-design-tokens": "5.2.0",
     "@nl-design-system-unstable/documentation": "1.9.0",
     "@nl-design-system-unstable/start-design-tokens": "6.1.0",
-    "@storybook/react-vite": "10.4.2",
     "@tabler/icons": "3.44.0",
     "@utrecht/checkbox-css": "3.0.1",
     "@utrecht/form-field-css": "3.0.1",
@@ -106,12 +107,12 @@
   },
   "devDependencies": {
     "@projectwallace/css-parser": "0.16.0",
+    "@storybook/react-vite": "10.4.2",
     "@types/dlv": "1.1.5",
     "@types/react": "19.2.17",
     "@types/react-dom": "19.2.3",
     "@vitest/browser-playwright": "4.1.8",
     "@vitest/coverage-v8": "4.1.8",
-    "glob": "13.0.6",
     "playwright": "1.60.0",
     "postcss-lit": "1.4.1",
     "rimraf": "6.1.3",
