+++ packages/theme-wizard-website/src/pages/components/[...slug]/tokens.astro
@@ -7,7 +7,7 @@ import { Traverse } from 'neotraverse/modern'
 import '@nl-design-system-candidate/paragraph-css/paragraph.css'
 import '@nl-design-system-candidate/link-css/link.css'
 import '@nl-design-system-community/clippy-components/clippy-story-preview';
-import { getStories } from '../../../../../theme-wizard-app/src/utils/csf-utils'
+import { getStories } from '@nl-design-system-community/theme-wizard-app'
 import '@nl-design-system-community/clippy-components/clippy-reset-theme'
 import './index.css';
